+++ src/map.rs
@@ -74,12 +74,12 @@ impl Widget for Map<'_, '_> {
     }
 }
 
-/// Position of the map's center. Initially, the map follows `my_position` argument which typically
+/// Position at the map's center. Initially, the map follows `my_position` argument which typically
 /// is meant to be fed by a GPS sensor or other geo-localization method. If user drags the map,
 /// it becomes "detached" and stays this way until [`MapMemory::center_mode`] is changed back to
 /// [`MapCenterMode::MyPosition`].
 #[derive(Clone, PartialEq)]
-pub enum MapCenterMode {
+pub enum Center {
     /// Center at `my_position` argument of the [`Map::new()`] function.
     MyPosition,
 
@@ -87,35 +87,36 @@ pub enum MapCenterMode {
     Exact(Position),
 }
 
-impl MapCenterMode {
+impl Center {
     fn screen_drag(&mut self, response: &Response, my_position: Position, zoom: u8) {
         if response.dragged_by(egui::PointerButton::Primary) {
             // We always end up in some exact, "detached" position, regardless of the current mode.
-            *self = MapCenterMode::Exact(screen_to_position(
+            *self = Center::Exact(screen_to_position(
                 self.position(my_position).project(zoom) - response.drag_delta(),
                 zoom,
             ));
         }
     }
 
+    /// Get the real position at the map's center.
     pub fn position(&self, my_position: Position) -> Position {
         match self {
-            MapCenterMode::MyPosition => my_position,
-            MapCenterMode::Exact(position) => *position,
+            Center::MyPosition => my_position,
+            Center::Exact(position) => *position,
         }
     }
 }
 
 /// State of the map widget which must persist between frames.
 pub struct MapMemory {
-    pub center_mode: MapCenterMode,
+    pub center_mode: Center,
     pub zoom: Zoom,
 }
 
 impl Default for MapMemory {
     fn default() -> Self {
         Self {
-            center_mode: MapCenterMode::MyPosition,
+            center_mode: Center::MyPosition,
             zoom: Default::default(),
         }
     }